+++ package.json
@@ -12,6 +12,7 @@
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.3.4",
     "react-scripts": "4.0.3",
     "web-vitals": "^1.0.1"
   },
