+++ src/components/app/App.js
@@ -1,43 +1,28 @@
-import { useState } from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import { MainPage, ComicsPage } from '../pages'
+
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import AppBanner from "../appBanner/AppBanner"
-import ComicsList from "../comicsList/ComicsList";
-import decoration from '../../resources/img/vision.png';
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+
 
 const App = () => {
 
-    const [selectedChar, setChar] = useState(null);
-
-    const onCharSelected = (id) => {
-        setChar(id);
-    }
-
     return (
-        <div className="app">
-            <AppHeader />
-            <main>
+        <Router>
+            <div className="app">
+                <AppHeader />
+                <main>
+                    <Switch>
+                        <Route exact path="/">
+                            <MainPage/>
 
-
-                {/* <ErrorBoundary>
-                    <RandomChar />
-                </ErrorBoundary>
-                <div className="char__content">
-                    <ErrorBoundary>
-                        <CharList onCharSelected={onCharSelected} />
-                    </ErrorBoundary>
-                    <ErrorBoundary>
-                        <CharInfo charId={selectedChar} />
-                    </ErrorBoundary>
-                </div> */}
-                <AppBanner />
-                <ComicsList />
-                <img className="bg-decoration" src={decoration} alt="vision" />
-            </main>
-        </div>
+                        </Route>
+                        <Route exact path="/comics">
+                            <ComicsPage/>
+                        </Route>
+                    </Switch>
+                </main>
+            </div>
+        </Router>
     )
 }
