+++ src/components/pages/ComicsPage.js
@@ -1,0 +1,13 @@
+import AppBanner from "../appBanner/AppBanner"
+import ComicsList from "../comicsList/ComicsList";
+
+const ComicsPage = () => {
+    return (
+        <>
+            <AppBanner />
+            <ComicsList />
+        </>
+    )
+}
+
+export default ComicsPage;
